readline!
I'm a little nervous about using this commit introducing rustyline,
because the other week when I tried to go further and implement
completions, I found out that the documentation is full of lies and
the crate appears to be unmaintained. (I wouldn't be so bothered by
one or the other of those, but the conjunction gives me pause.) But
then I tried to use what appears to be the most popular linenoise
wrapper, and up-arrow history didn't seem to work (even though the
method names in the example code made it look like it would work), and
there's an open issue from 2016 (octplane/rust-linenoise#12) that
looks pretty damning.

So ... let's just let this be on master and maybe swap it out later if
we want completions and really can't make them work in
rustyline. Consider it my karmic punishment for my delinquent
co-maintainership of rusty-machine.